Moved get_case_errors and validate_detail_columns
