+++ corehq/apps/app_manager/helpers/validators.py
@@ -356,6 +356,33 @@ class ModuleBaseValidator(object):
 
         return errors
 
+    def validate_detail_columns(self, columns):
+        from corehq.apps.app_manager.suite_xml.const import FIELD_TYPE_LOCATION
+        from corehq.apps.locations.util import parent_child
+        from corehq.apps.locations.fixtures import should_sync_hierarchical_fixture
+
+        hierarchy = None
+        for column in columns:
+            if column.field_type == FIELD_TYPE_LOCATION:
+                domain = self.module.get_app().domain
+                project = Domain.get_by_name(domain)
+                try:
+                    if not should_sync_hierarchical_fixture(project, self.module.get_app()):
+                        # discontinued feature on moving to flat fixture format
+                        raise LocationXpathValidationError(
+                            _('That format is no longer supported. To reference the location hierarchy you need to'
+                              ' use the "Custom Calculations in Case List" feature preview. For more information '
+                              'see: https://confluence.dimagi.com/pages/viewpage.action?pageId=38276915'))
+                    hierarchy = hierarchy or parent_child(domain)
+                    LocationXpath('').validate(column.field_property, hierarchy)
+                except LocationXpathValidationError as e:
+                    yield {
+                        'type': 'invalid location xpath',
+                        'details': six.text_type(e),
+                        'module': self.module.get_module_info(),
+                        'column': column,
+                    }
+
 
 class ModuleDetailValidatorMixin(object):
     '''
@@ -400,6 +427,32 @@ class ModuleDetailValidatorMixin(object):
                             'reason': _('A case property must be assigned to the "{}" tile field.'.format(field))
                         })
         return errors
+
+    def get_case_errors(self, needs_case_type, needs_case_detail, needs_referral_detail=False):
+        module_info = self.module.get_module_info()
+
+        if needs_case_type and not self.module.case_type:
+            yield {
+                'type': 'no case type',
+                'module': module_info,
+            }
+
+        if needs_case_detail:
+            if not self.module.case_details.short.columns:
+                yield {
+                    'type': 'no case detail',
+                    'module': module_info,
+                }
+            columns = self.module.case_details.short.columns + self.module.case_details.long.columns
+            errors = self.validate_detail_columns(columns)
+            for error in errors:
+                yield error
+
+        if needs_referral_detail and not self.module.ref_details.short.columns:
+            yield {
+                'type': 'no ref detail',
+                'module': module_info,
+            }
 
     def _validate_detail_screen_field(self, field):
         # If you change here, also change here:
@@ -511,6 +564,37 @@ class AdvancedModuleValidator(ModuleBaseValidator):
                     })
 
         return errors
+
+    def get_case_errors(self, needs_case_type, needs_case_detail, needs_referral_detail=False):
+        module_info = self.module.get_module_info()
+
+        if needs_case_type and not self.module.case_type:
+            yield {
+                'type': 'no case type',
+                'module': module_info,
+            }
+
+        if needs_case_detail:
+            if not self.module.case_details.short.columns:
+                yield {
+                    'type': 'no case detail',
+                    'module': module_info,
+                }
+            if self.module.get_app().commtrack_enabled and not self.module.product_details.short.columns:
+                for form in self.module.forms:
+                    if self.module.case_list.show or \
+                            any(action.show_product_stock for action in form.actions.load_update_cases):
+                        yield {
+                            'type': 'no product detail',
+                            'module': module_info,
+                        }
+                        break
+            columns = self.module.case_details.short.columns + self.module.case_details.long.columns
+            if self.module.get_app().commtrack_enabled:
+                columns += self.module.product_details.short.columns
+            errors = self.validate_detail_columns(columns)
+            for error in errors:
+                yield error
 
 
 class ReportModuleValidator(ModuleBaseValidator):
